fixup! Fixup seeding for portability to web

+++ include/dish2/introspection/get_endeavor.hpp
@@ -12,9 +12,9 @@
 
 namespace dish2 {
 
-emp::optional<size_t> get_endeavor() {
+emp::optional<uint32_t> get_endeavor() {
 
-  if ( cfg.SERIES() != std::numeric_limits<size_t>::max() )
+  if ( cfg.SERIES() != std::numeric_limits<uint32_t>::max() )
     return cfg.SERIES() % 1000;
   else if ( std::getenv("DISHTINY_ENDEAVOR") )
     return uitsl::stoszt( std::getenv("DISHTINY_ENDEAVOR") );
